(reciever_test_2.0 updates) This commit basically takes what i have been working on and adds some commented psudocode to where and what i would need to include to take what we currently have to a full on PID loop controller on the side of the drone.
The transmitter test is basically updating the data type being sent over the radio protocols so that the number of bits interpreted
and read on the reciever sent from the controller is the same. Switching from the arduino nano reciever test to the teensy on the side of the
reciever was getting some errors by reading in too many bits and causing problems.

+++ Transmitter + Reciever Data/transmitter_test_2.0/transmitter_test_2.0.ino
@@ -24,10 +24,10 @@ const byte address[6] = "00001"; // Must match the receiver [cite: 19]
 
 // Define a structure to hold all 4 joystick values
 struct ControlData {
-  int throttle;
-  int yaw;
-  int pitch;
-  int roll;
+  int16_t throttle;
+  int16_t yaw;
+  int16_t pitch;
+  int16_t roll;
 };
 
 ControlData data; // Create an instance of our data structure
